Title: Added basic envs for Mobile Robot use-case
Body: Added a maze and empty env for a mobile robot use case using the
kobuki robot. The maze env is still going through revisions as the
training is taking too long and possibly the size of the maze is going
to be reduced. The training of the empty env is already completed and
the logs were generated for both TD3 and SAC.

Authored-by: Jose Fajardo <jmfajardod@unal.edu.co>