Title: Added KobukiEnv with dynamic obstacles
Body: Created three differents envs with the kobuki robot where the
goal is to reach a random position. In the env there are eight boxes
with random orientation and velocity. Created a test file but
functionality to the library is needed so the library can use the
load and predict functions of SB3.

Authored-by: Jose Fajardo <jmfajardod@unal.edu.co>